[cli] rename to @build-tracker/utility, expose methods
