+++ src/utility/index.js
@@ -1,0 +1,4 @@
+/**
+ * Copyright (c) 2019 Paul Armstrong
+ */
+export * from './dist';
