+++ src/utility/index.ts
@@ -1,3 +1,4 @@
 /**
  * Copyright (c) 2019 Paul Armstrong
  */
+export * from './src';
