+++ src/utility/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@build-tracker/cli",
+  "name": "@build-tracker/utility",
   "version": "1.0.0-beta.9",
   "description": "Build Tracker tool for reading build artifact sizes and uploading to your server",
   "author": "Paul Armstrong <paul@spaceyak.com>",
@@ -9,6 +9,7 @@
     "bt-cli": "./dist/bin.js"
   },
   "main": "dist",
+  "module": "./",
   "scripts": {
     "build": "tsc",
     "clean": "rimraf dist",
