+++ src/utility/src/commands/__tests__/create-build.test.ts
@@ -1,0 +1,53 @@
+/**
+ * Copyright (c) 2019 Paul Armstrong
+ */
+import * as brotliSize from 'brotli-size';
+import * as Command from '../create-build';
+import * as Git from '../../modules/git';
+import path from 'path';
+import yargs from 'yargs';
+
+const config = path.join(
+  path.dirname(require.resolve('@build-tracker/fixtures')),
+  'cli-configs/rc/.build-trackerrc.js'
+);
+
+describe('create-build', () => {
+  describe('builder', () => {
+    test('defaults config', () => {
+      const args = Command.builder(yargs([]));
+      expect(args.argv).toEqual({
+        $0: expect.any(String),
+        _: [],
+        o: true,
+        out: true,
+        'skip-dirty-check': false,
+        skipDirtyCheck: false
+      });
+    });
+  });
+
+  describe('handler', () => {
+    let writeSpy;
+    beforeEach(() => {
+      writeSpy = jest.spyOn(process.stdout, 'write').mockImplementationOnce(() => true);
+      jest.spyOn(Git, 'isDirty').mockImplementation(() => Promise.resolve(false));
+      jest.spyOn(Git, 'getDefaultBranch').mockImplementation(() => Promise.resolve('master'));
+      jest.spyOn(Git, 'getMergeBase').mockImplementation(() => Promise.resolve('1234567'));
+      jest.spyOn(Git, 'getParentRevision').mockImplementation(() => Promise.resolve('7654321'));
+      jest.spyOn(Git, 'getCurrentRevision').mockImplementation(() => Promise.resolve('abcdefg'));
+      jest.spyOn(Git, 'getBranch').mockImplementation(() => Promise.resolve('master'));
+      jest
+        .spyOn(Git, 'getRevisionDetails')
+        .mockImplementation(() => Promise.resolve({ timestamp: 1234567890, name: 'Jimmy', subject: 'tacos' }));
+      jest.spyOn(brotliSize, 'sync').mockImplementation(() => 49);
+    });
+
+    test('writes the artifact stats to stdout', async () => {
+      await expect(Command.handler({ config, out: true, 'skip-dirty-check': true })).resolves.toEqual(
+        expect.any(Object)
+      );
+      expect(writeSpy).toHaveBeenCalledWith(expect.stringMatching('\\"parentRevision\\": \\"7654321\\"'));
+    });
+  });
+});
