+++ src/utility/src/commands/stat-artifacts.ts
@@ -3,9 +3,7 @@
  */
 import { Argv } from 'yargs';
 import getConfig from '../modules/config';
-import glob from 'glob';
-import path from 'path';
-import readfile from '../modules/readfile';
+import statArtifacts, { Stat } from '../modules/stat-artifacts';
 
 export const command = 'stat-artifacts';
 
@@ -16,16 +14,7 @@ interface Args {
   out: boolean;
 }
 
-interface Stat {
-  hash: string;
-  stat: number;
-  gzip: number;
-  brotli: number;
-}
-
 const group = 'Stat artifacts';
-
-const defaultNameMapper = (fileName: string): string => fileName;
 
 export const builder = (yargs): Argv<Args> =>
   yargs
@@ -44,18 +33,10 @@ export const builder = (yargs): Argv<Args> =>
       type: 'boolean'
     });
 
-export const handler = async (args: Args): Promise<{ artifacts: Map<string, Stat> }> => {
-  const { artifacts: artifactGlobs, baseDir, cwd, getFilenameHash, nameMapper = defaultNameMapper } = await getConfig(
-    args.config
-  );
+export const handler = async (args: Args): Promise<Map<string, Stat>> => {
+  const config = await getConfig(args.config);
 
-  const artifacts = new Map();
-  artifactGlobs.forEach(fileGlob => {
-    glob.sync(path.resolve(cwd, fileGlob), { nodir: true }).forEach(filePath => {
-      const sizes = readfile(filePath, getFilenameHash);
-      artifacts.set(nameMapper(path.relative(baseDir, filePath).replace(`.${sizes.hash}`, '')), sizes);
-    });
-  }, []);
+  const artifacts = statArtifacts(config);
 
   if (args.out) {
     const fileOut = Array.from(artifacts).reduce((memo, [artifactName, stat]) => {
@@ -66,7 +47,5 @@ export const handler = async (args: Args): Promise<{ artifacts: Map<string, Stat
     process.stdout.write(JSON.stringify(fileOut, null, 2));
   }
 
-  return Promise.resolve({
-    artifacts
-  });
+  return Promise.resolve(artifacts);
 };
