+++ src/utility/src/commands/version.ts
@@ -10,5 +10,5 @@ export const description = 'Print the Build Tracker version';
 export const builder = (yargs): Argv<{}> => yargs.usage(`Usage: $0 ${command}`);
 
 export const handler = (): void => {
-  process.stdout.write(`${require('@build-tracker/cli/package.json').version}\n`);
+  process.stdout.write(`${require('@build-tracker/utility/package.json').version}\n`);
 };
